♻️ :: (#72) code enum과 converter 분리
