+++ application-domain/src/main/java/io/github/v1serviceapplication/stay/api/ApplyStay.java
@@ -1,6 +1,8 @@
 package io.github.v1serviceapplication.stay.api;
 
 
+import io.github.v1serviceapplication.stay.code.StayStatusCode;
+
 public interface ApplyStay {
-    void applyStay(String status);
+    void applyStay(StayStatusCode status);
 }
