+++ application-domain/src/main/java/io/github/v1serviceapplication/stay/spi/StayRepositorySpi.java
@@ -1,8 +1,9 @@
 package io.github.v1serviceapplication.stay.spi;
 
 import io.github.v1serviceapplication.annotation.Spi;
+import io.github.v1serviceapplication.stay.code.StayStatusCode;
 
 @Spi
 public interface StayRepositorySpi {
-    void applyStay(String status);
+    void applyStay(StayStatusCode status);
 }
