+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stay/domain/StayEntity.java
@@ -1,7 +1,8 @@
 package io.github.v1serviceapplication.domain.stay.domain;
 
+import io.github.v1serviceapplication.domain.stay.converter.StayStatusCodeConverter;
 import io.github.v1serviceapplication.global.entity.BaseEntity;
-import io.github.v1serviceapplication.domain.stay.code.StayStatusCode;
+import io.github.v1serviceapplication.stay.code.StayStatusCode;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
@@ -17,6 +18,6 @@ import javax.validation.constraints.NotNull;
 public class StayEntity extends BaseEntity {
     @NotNull
     @Column(columnDefinition = "CHAR(6)")
-    @Convert(converter = StayStatusCode.StayStatusCodeConverter.class)
+    @Convert(converter = StayStatusCodeConverter.class)
     private StayStatusCode code;
 }
